Update guess_the_number_game.py
change float to int

+++ guess_the_number_game.py
@@ -3,7 +3,7 @@ import random
 num = random.randint(1,100)
 while True:
 	numa = input('猜猜数字是什么：')
-	numa = float(numa)
+	numa = int(numa)
 	if numa == num:
 		print('你猜对啦！')
 		break
